Rework menu, to allow for Touch Controls
The main part of this commit is adding touch controls to all the menus. That includes the Main Menu, Options Menu, Stage Select, Pause Menu, and Results Menu (doesn't include Special Stage Pause Menu because that's a wip). There's also small cleanup around the place too, like the background of the menus being changed to use tiles rather than a few hundred or so sprites.